+++ src/main/kotlin/graphql/operations/Queries.kt
@@ -261,13 +261,3 @@ fun searchPublicChats(env: DataFetchingEnvironment): List<GroupChatDto> {
         messagesPagination = BackwardPagination(last = 0)
     ).map { GroupChatDto(it.id) }
 }
-
-fun isBlocked(env: DataFetchingEnvironment): Boolean {
-    env.verifyAuth()
-    return BlockedUsers.exists(env.userId!!, env.getArgument("id"))
-}
-
-fun isContact(env: DataFetchingEnvironment): Boolean {
-    env.verifyAuth()
-    return env.getArgument("id") in Contacts.readIdList(env.userId!!)
-}